Ajout statut et protection des routes

+++ app/Http/Controllers/ToDoController.php
@@ -16,7 +16,9 @@ class ToDoController extends Controller
      */
     public function index()
     {
-        $todo = ToDo::get();
+        $user = auth()->user();
+        $todo = ToDo::where('user_id', $user->id)->get();
+
         return Inertia::render('FrontEnd/Todo/Index', [
             'todo' => $todo
         ]);
@@ -37,12 +39,13 @@ class ToDoController extends Controller
     {
         $request->validate([
             'title' => 'required|string|max:255',
-            'more_info' => 'string|max:255'
+            'more_info' => 'required|string|max:255',
         ]);
 
         ToDo::create([
             'title' => $request->title,
             'more_info' => $request->more_info,
+            'user_id' => auth()->id(),
         ]);
 
         return redirect()->to('/todo')->with('message','Tache ajoutée à la liste');
@@ -75,12 +78,14 @@ class ToDoController extends Controller
     {
         $request->validate([
             'title' => 'required|string|max:255',
-            'more_info' => 'string|max:255'
+            'more_info' => 'string|max:255',
+            'statut' => 'required|string'
         ]);
 
         $todo->update([
             'title' => $request->title,
             'more_info' => $request->more_info,
+            'statut' => $request->statut,
         ]);
 
         return redirect()->to('/todo')->with('message','Tache mise à jour');
@@ -89,8 +94,12 @@ class ToDoController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(ToDo $todo)
+    public function destroy(Request $request, ToDo $todo)
     {
+        if ($todo->statut !== 'termine') {
+            return redirect()->back()->with('message', 'Seules les tâches terminées peuvent être supprimées.');
+        }
+
         $todo->delete();
         return redirect()->to('/todo')->with('message','Tache supprimer de la liste');
